backend/requirements.txt
Django==6.0.1
asgiref==3.11.0
gunicorn==24.1.1

djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
django-filter==25.2

django-allauth==65.14.0
PyJWT==2.10.1
oauthlib==3.3.1

psycopg2-binary==2.9.11
redis==7.1.0
django-redis==6.0.0
sqlparse==0.5.5

Pillow==12.1.0
pilkit==3.0
django-imagekit==6.0.0

cryptography==46.0.3
cffi==2.0.0
pycparser==3.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3
requests==2.32.5

jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==25.4.0
packaging==26.0
typing_extensions==4.15.0
tzdata==2025.3
phonenumbers==9.0.22
inflection==0.5.1
uritemplate==4.2.0
django-appconf==1.2.0
django-environ==0.12.0
django-cors-headers==4.9.0
PyYAML==6.0.3
Pygments==2.19.2

pytest==9.0.2
pytest-django==4.11.1
pluggy==1.6.0
iniconfig==2.3.0
